[rasterize] fix index out of bounds panic in signed_difference_line
In certain rare cases x_next can be set to a value slightly greater than
the width of the image. This occurs due to floating point rounding
errors when the line touches or crosses the positive x boundary of
the image. Specifically the calculation `x + dxdy * dy` introduces
the error, then the ceil() of that value is taken before it is used to
index into the array of pixels.